Create desolate_earth.png
Create desolate_earth.png from Earth.jpg. I indexed the Earth image (set to 100 colors) and also indexed an image of dry earth in order to get the colors to make the Earth look like a large desert. I also added the gray to look either like smoke or burnt ground. The red and orange throughout are to represent fire