Update progress_preview1.png
Update progress_preview1.png to show the progress I made before the first preview. This is where I indexed all of the colors of the image of Earth to make the "desolate" looking Earth. It should be noted this was taken at a later date, but still shows the progress my project went through