Update progress_preview2.png
Update progress_preview2.png to show the progress I made before the first preview. This image shows the added slice of Earth with the colors vibrancy increased. It should be noted this was taken at a later date, though it still shows the progress that my project followed throughout it's creation.