Create progress_preview10.png
Create progress_preview10.png, which was the second to last iteration for my project. Here I removed the reference to climate change in the bottom text to shift the focus fully on the Doomsday Clock. I didn't stick with this final version because it didn't give enough information regarding the Doomsday Clock. The final version has a link for more information in the bottom text.